feat: rename project-artifacts
